feat!: integrates validator node exits, updates registration (#1405)
Description
---
feat!: integrates validator node exits
fix!: fixes validator node signature replay issue by binding the
signature to an epoch
feat!: add JSON-RPC call to prompt the validator to create a signed
registration transaction file
feat: implement transaction file submitter in swarm
feat(db-inspector): add custom endpoint for state_transitions column
family
feat(swarm): add `log_to_file` config option (defaults to false, due to
potentially high disk space usage)

Motivation and Context
---
Ref https://github.com/tari-project/tari/pull/6955
Closes #1398 

How Has This Been Tested?
---
Manually 

What process can a PR reviewer use to test or verify this change?
---
Click 'Submit Exit' in swarm on a VN and mine the UTXO in - observe the
VN transitions to Idle

Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [ ] Other - Please specify

BREAKING CHANGE: additional layer-one transaction types that require an
updated base layer

+++ applications/tari_swarm_daemon/src/process_manager/instances/manager.rs
@@ -308,8 +308,8 @@ impl InstanceManager {
         self.instances().count()
     }
 
-    pub fn num_validator_nodes(&self) -> u64 {
-        self.validator_nodes.len() as u64
+    pub fn num_validator_nodes(&self) -> usize {
+        self.validator_nodes.len()
     }
 
     pub fn validator_nodes_mut(&mut self) -> impl Iterator<Item = &mut ValidatorNodeProcess> + Sized {
